_handle_match: use non-capturing regexes
Instead of using the usual r'(foo)' capture groups for "foo", use
non-capturing groups like r'(?:foo)' for the data that we don't care
about.

This allows us to simplify the code and avoid using lists-of-lists,
since re.findall() will only return the data that we will examine futher
on.

This will also make it easier to support multiple pp_pattern matches in
a single match() operation (a separate change, not part of this commit).